Lighter red notice; changed default inhalation rates
Also, I went back to look at inhalation rates. This was was something I had forgotten to as I  inappropriate used moderate intensity. So. based on he ref, probably best to use mean-95th percentile range for students age 16 to <21 for sedentary/passive activity and for faculty age 41 < 51 mean passive activity to 80% or so light activity.

So an you also please change default inhalation rates as:
0.005 - 0.007 for students
0.005 - 0.01 faculty